requirements.txt
beautifulsoup4==4.4.1
coreapi==2.0.8
Django==1.10.1
django-cors-headers==1.3.0
django-environ==0.4.0
django-filter==0.15.3
django-rest-swagger==2.1.0
djangorestframework==3.5.1
feedparser==5.2.1
httplib2==0.9.2
itypes==1.1.0
oauth2==1.9.0.post1
openapi-codec==1.1.7
Pillow==3.3.1
psycopg2==2.6.2
requests==2.11.1
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
yelp==1.0.2
git+https://github.com/senko/python-video-converter.git@0cad6b0ec16e345da56f36f06b7a19bc5278d117#egg=VideoConverter
https://github.com/slimkrazy/python-google-places/zipball/master